WASM bindings moved to the project root and simplified (one handler for all modules unlike multiple nested decorators and dispatchers)
