+++ pkg/coins.go
@@ -1,9 +1,6 @@
 package pkg
 
 import (
-	"fmt"
-
-	wasmVmTypes "github.com/CosmWasm/wasmvm/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -25,31 +22,4 @@ func SplitCoins(coins sdk.Coins, ratio sdk.Dec) (stack1, stack2 sdk.Coins) {
 	}
 
 	return
-}
-
-// WasmCoinToSDK converts wasmVmTypes.Coin to sdk.Coin.
-func WasmCoinToSDK(coin wasmVmTypes.Coin) (sdk.Coin, error) {
-	amount, ok := sdk.NewIntFromString(coin.Amount)
-	if !ok {
-		return sdk.Coin{}, fmt.Errorf("invalid amount: %s", coin.Amount)
-	}
-
-	return sdk.Coin{
-		Denom:  coin.Denom,
-		Amount: amount,
-	}, nil
-}
-
-// WasmCoinsToSDK converts wasmVmTypes.Coins to sdk.Coins.
-func WasmCoinsToSDK(coins wasmVmTypes.Coins) (sdk.Coins, error) {
-	result := make(sdk.Coins, 0, len(coins))
-	for _, coin := range coins {
-		coinSDK, err := WasmCoinToSDK(coin)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, coinSDK)
-	}
-
-	return result, nil
 }
