+++ x/rewards/spec/08_wasm_bindings.md
@@ -8,16 +8,36 @@ Section describes interaction with the module by a contract using CosmWasm custo
 
 ## Custom query
 
-[The custom query structure](../wasmbinding/types/query.go#L19) is used to query the `x/rewards` specific data by a contract.
+[The custom query structure](../../../wasmbinding/types/query.go#L10) is used to query a specific module data by a contract.
+
+This query is expected to fail if:
+
+* Query has no sub-query specified (`rewards` field is not defined);
+* Query has more than one sub-query specified;
+
+## Custom message
+
+[The custom message structure](../../../wasmbinding/types/msg.go#L10) is used to send a module specific state change message by a contract.
+
+This message is expected to fail if:
+
+* Message has no sub-message specified (`rewards` field is not defined);
+* Message has more than one sub-message specified;
+
+## Rewards module bindings
+
+### Queries
+
+[The sub-query structure](../../../wasmbinding/rewards/types/query.go#L8) is used to query the `x/rewards` module specific data.
 
 This query is expected to fail if:
 
 * Query has no request specified (`metadata` and `rewards_records` fields are not defined);
 * Query has more than one request specified;
 
-### Metadata
+#### Metadata
 
-The [metadata](../wasmbinding/types/query.go#L37) request returns a contract metadata state.
+The [metadata](../../../wasmbinding/rewards/types/query_metadata.go#L12) request returns a contract metadata state.
 A contract can query its own or any other contract's metadata.
 
 Query example:
@@ -41,18 +61,18 @@ Example response:
 }
 ```
 
-### Rewards records
+#### Rewards records
 
-The [rewards_records](../wasmbinding/types/query.go#L53) request returns the paginated list of `RewardsRecord` objects credited to an account address.
-A [RewardsRecord](../wasmbinding/types/query.go#L70) entry contains a portion of credited rewards by a specific contract at a block height.
+The [rewards_records](../../../wasmbinding/rewards/types/query_records.go#L17) request returns the paginated list of `RewardsRecord` objects credited to an account address.
+A [RewardsRecord](../../../wasmbinding/rewards/types/query_records.go#L35) entry contains a portion of credited rewards by a specific contract at a block height.
 A contract can query any account address rewards state.
 
 This query is paginated to limit the size of the response.
-Refer to the [PageRequest](../wasmbinding/types/pagination.go#L8) structure description to learn more about the pagination options.
-The query response contains the [PageResponse](../wasmbinding/types/pagination.go#L28) structure that should be used to query the next page.
+Refer to the [PageRequest](../../../wasmbinding/pkg/pagination.go#L8) structure description to learn more about the pagination options.
+The query response contains the [PageResponse](../../../wasmbinding/pkg/pagination.go#L28) structure that should be used to query the next page.
 
 > The maximum page limit is bounded by the `MaxWithdrawRecords` parameter.
-> 
+>
 > If the page limit is not set, the default value is `MaxWithdrawRecords`.
 
 Query example:
@@ -94,18 +114,18 @@ Example response:
 }
 ```
 
-## Custom message
+### Messages
 
-[The custom message structure](../wasmbinding/types/msg.go#L15) is used to send the `x/rewards` state change messages by a contract.
+[The sub-message structure](../../../wasmbinding/rewards/types/msg.go#L8) is used to send the `x/rewards` module specific state change message.
 
 This message is expected to fail if:
 
-* Message has no sub-message specified (`update_metadata` and `withdraw_rewards` fields are not defined);
-* Message has more than one sub-message specified;
+* Message has no operations specified (`update_metadata` and `withdraw_rewards` fields are not defined);
+* Message has more than one operation specified;
 
-### Update metadata
+#### Update metadata
 
-The [update_metadata](../wasmbinding/types/msg.go#L34) request is used to update an existing contract metadata.
+The [update_metadata](../../../wasmbinding/rewards/types/msg_metadata.go#L12) request is used to update an existing contract metadata.
 
 Message example (CosmWasm's `CosmosMsg`):
 
@@ -136,9 +156,9 @@ This sub-message is expected to fail if:
 * No fields to update were set (`owner_address` and `rewards_address` are empty);
 * The contract address is not set as the metadata's `owner_address` (request is unauthorized);
 
-### Withdraw rewards
+#### Withdraw rewards
 
-The [withdraw_rewards](../wasmbinding/types/msg.go#L44) request is used to withdraw the current credited to a contract address reward tokens.
+The [withdraw_rewards](../../../wasmbinding/rewards/types/msg_withdraw.go#L12) request is used to withdraw the current credited to a contract address reward tokens.
 
 > Contract address is used as the `rewards_address` for this sub-message: a contract can only request withdrawal of funds, credited for his own address.
 
@@ -169,7 +189,7 @@ Message example (CosmWasm's `CosmosMsg`):
 }
 ```
 
-Sub-message returns the [response](../wasmbinding/types/msg.go#L55) that can be handled with the *Reply* CosmWasm functionality.
+Sub-message returns the [response](../../../wasmbinding/rewards/types/msg_withdraw.go#L23) that can be handled with the *Reply* CosmWasm functionality.
 
 Response example:
